+++ packages/@cldcvr/flow-form-builder/src/components/f-form-builder/__snapshots__/f-form-builder.test.snap.js
@@ -24,6 +24,7 @@ snapshots["f-form-builder should render object field with all default properties
     label="{&quot;title&quot;:&quot;Object field form&quot;,&quot;description&quot;:&quot;showing object field&quot;}"
   >
     <f-input
+      autocomplete="off"
       category="fill"
       clear=""
       data-qa-element-id=""
@@ -54,6 +55,7 @@ snapshots["f-form-builder should render object field with all default properties
       </f-div>
     </f-input>
     <f-input
+      autocomplete="off"
       category="fill"
       clear=""
       data-qa-element-id=""
@@ -111,6 +113,7 @@ snapshots["f-form-builder should emit state-change with validation messages"] =
     label="{&quot;title&quot;:&quot;Object field form&quot;,&quot;description&quot;:&quot;showing object field&quot;}"
   >
     <f-input
+      autocomplete="off"
       category="fill"
       clear=""
       data-qa-element-id=""
@@ -141,6 +144,7 @@ snapshots["f-form-builder should emit state-change with validation messages"] =
       </f-div>
     </f-input>
     <f-input
+      autocomplete="off"
       category="fill"
       clear=""
       data-qa-element-id=""
@@ -219,6 +223,7 @@ snapshots["f-form-builder should display array fields"] =
       width="fill-container"
     >
       <f-input
+        autocomplete="off"
         category="fill"
         clear=""
         data-qa-element-id=""
@@ -256,6 +261,7 @@ snapshots["f-form-builder should display array fields"] =
       width="fill-container"
     >
       <f-input
+        autocomplete="off"
         category="fill"
         clear=""
         data-qa-element-id=""
@@ -329,6 +335,7 @@ snapshots["f-form-builder should add/delete field in array"] =
       width="fill-container"
     >
       <f-input
+        autocomplete="off"
         category="fill"
         clear=""
         data-qa-element-id=""
@@ -366,6 +373,7 @@ snapshots["f-form-builder should add/delete field in array"] =
       width="fill-container"
     >
       <f-input
+        autocomplete="off"
         category="fill"
         clear=""
         data-qa-element-id=""
@@ -459,6 +467,7 @@ snapshots["f-form-builder should test snapshot of all fields"] =
     >
     </f-divider>
     <f-input
+      autocomplete="off"
       category="fill"
       clear=""
       data-qa-element-id=""
@@ -515,6 +524,7 @@ snapshots["f-form-builder should test snapshot of all fields"] =
     >
     </f-divider>
     <f-input
+      autocomplete="off"
       category="fill"
       clear=""
       data-hidden="false"