+++ packages/@cldcvr/flow-form-builder/src/index.ts
@@ -5,3 +5,5 @@ export * from "./components/f-radio-group/f-radio-group";
 export * from "./components/f-form-array/f-form-array";
 export * from "./components/f-form-object/f-form-object";
 export * from "./components/f-field-separator/f-field-separator";
+
+export { html } from "lit";